issue6,7 : 리뷰 수정, 삭제

+++ src/main/java/com/smell/moviereview/service/ReviewServiceImpl.java
@@ -3,6 +3,7 @@ package com.smell.moviereview.service;
 import com.smell.moviereview.dao.ReviewDAO;
 import com.smell.moviereview.service.impl.ReviewService;
 import com.smell.moviereview.vo.ReviewVO;
+import com.smell.moviereview.vo.ReviewerVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -21,13 +22,29 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    public void modifyReview(ReviewVO review) {
-        reviewDao.modify(review);
+    public void modifyReview(int id, ReviewVO modifyReview) {
+        ReviewVO review = reviewDao.selectOneReview(id);
+
+        if(review != null){
+            review.setRating(modifyReview.getRating());
+            review.setTitle(modifyReview.getTitle());
+            review.setDirector(modifyReview.getDirector());
+            review.setActor(modifyReview.getActor());
+            review.setContents(modifyReview.getContents());
+            reviewDao.modify(review);
+        }else{
+            throw new IllegalStateException("해당 리뷰는 존재하지 않습니다.");
+        }
+
     }
 
     @Override
     public void removeReview(int id) {
-        reviewDao.remove(id);
+        if(reviewDao.selectOneReview(id) != null){
+            reviewDao.remove(id);
+        }else{
+            throw new IllegalStateException("이미 삭제되었습니다.");
+        }
     }
 
     @Override
